logpipe: add support for drop-unmatched flag
This patch adds a new flag to log statements, "drop-unmatched" will cause
messages to be dropped along a specific log path, should it not match the
filter or be dropped by a parser.  Normally in such cases, syslog-ng would
continue to process the message along alternative paths.

This is basically saying that a filter/parse error is final and is to be used
by our if statement.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>